requirements-dev.txt
# Development dependencies
-r requirements.txt

# Type checking
django-stubs>=4.2.0
djangorestframework-stubs>=3.16.0
mypy>=1.7.0

# Testing
pytest>=7.4.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
factory-boy>=3.3.0

# Code formatting
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0

# Pre-commit hooks
pre-commit>=3.5.0
